% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invlogit.R
\name{invlogit}
\alias{invlogit}
\title{Inverse Logit Function}
\usage{
invlogit(x)
}
\arguments{
\item{x}{numeric on the real line}
}
\value{
numeric between 0 and 1
}
\description{
This function maps a number x on the real line to invlogit(x) = 1/(1+exp(-x)
}
\examples{
logit(-0.8472979)
0.3
}
\keyword{invlogit}
\keyword{logit,}
