% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repmat.R
\name{repmat}
\alias{repmat}
\title{Rep Mat Function}
\usage{
repmat(X, m, n)
}
\arguments{
\item{X}{a numeric matrix}

\item{m}{number of row-wise replications}

\item{n}{number of column-wise replications}
}
\value{
numeric matrix with X replicated n-times across and m-times down
}
\description{
This function takes a matrix X and replicates it n times column-wise and m times row-wise, similarly to repmat in Matlab
}
\examples{
set.seed(1)
x = matrix(rnorm(6),2,3)
> x
         [,1]       [,2]       [,3]     
[1,] -0.6264538 -0.8356286  0.3295078
[2,]  0.1836433  1.5952808 -0.8204684
repmat(x,3,2)
           [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
[1,] -0.6264538 -0.8356286  0.3295078 -0.6264538 -0.8356286  0.3295078
[2,]  0.1836433  1.5952808 -0.8204684  0.1836433  1.5952808 -0.8204684
[3,] -0.6264538 -0.8356286  0.3295078 -0.6264538 -0.8356286  0.3295078
[4,]  0.1836433  1.5952808 -0.8204684  0.1836433  1.5952808 -0.8204684
[5,] -0.6264538 -0.8356286  0.3295078 -0.6264538 -0.8356286  0.3295078
[6,]  0.1836433  1.5952808 -0.8204684  0.1836433  1.5952808 -0.8204684
}
\keyword{repmat}
