% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softmax.R
\name{softmax}
\alias{softmax}
\title{Softmax Function}
\usage{
softmax(x)
}
\arguments{
\item{x}{numeric vector of real numbers}
}
\value{
numeric vector of proportions between 0 and 1 with a sum of 1
}
\description{
This function maps a numeric vector of real numbers to a vector of proportions, between 0 and 1, summing to 1.
}
\examples{
x = c(5,6,1,10)
softmax(x)
0.006572472 0.017865831 0.000120379 0.975441318
}
\keyword{softmax}
