% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorize.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize Function}
\usage{
winsorize(x, fraction)
}
\arguments{
\item{x}{numeric vector of real numbers}
}
\value{
numeric vector of x with all elements below the franctionth percentile truncated to the limiting value from the top and bottom
}
\description{
This function truncates vector to the fractionth-ile elements from the top and the bottom
}
\examples{
set.seed(1)
x <- rnorm(10)
x <- sort(x)
x
[1] -0.8356286 -0.8204684 -0.6264538 -0.3053884  0.1836433  0.3295078  0.4874291  0.5757814  0.7383247  1.5952808
winsorize(x,0.2)
[1] -0.6652567 -0.6652567 -0.6264538 -0.3053884  0.1836433  0.3295078  0.4874291  0.5757814  0.6082900  0.6082900
}
\keyword{winsorize}
